requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
jinja2==3.1.3
python-dotenv==1.0.1
opencv-python==4.9.0.80
numpy==1.26.4
aiofiles==23.2.1
depthai==2.24.0
pydantic==2.6.3
pydantic-settings==2.2.1